moved deprecated commands to own file
